add calculations

+++ sco.climb.game-dashboard/src/main/java/it/smartcommunitylab/climb/gamification/dashboard/model/CalendarDay.java
@@ -15,7 +15,7 @@ public class CalendarDay extends BaseObject {
 	private String meteo;
 
 	private Integer duration;
-	private Integer distance;
+	private Double distance;
 
 	private int EActive=0;
 	private int VActive=0;
@@ -82,9 +82,9 @@ public class CalendarDay extends BaseObject {
 
 	public void setDuration(Integer duration) {this.duration = duration;}
 
-	public Integer getDistance() {return distance;}
+	public Double getDistance() {return distance;}
 
-	public void setDistance(Integer distance) {this.distance = distance;}
+	public void setDistance(Double distance) {this.distance = distance;}
 
 
 	public int getEActive() {
